[TASK] improve delete message for tags/fieldcollections/objectbricks/… (#12102)
* [TASK] improve delete message for tags/fieldcollections/objectbricks/users and add confirmdialog to thumbnails/customreports

Partly resolves: #6259

* [TASK] add confirm dialogs to classificationstore delete operations (key/relation/collection/group)

Relates: #6259

* [TASK] add confirm dialogs videothumbnail delete operation

Relates: #6259

* [TASK] add confirm dialogs to targeting delete operations

Relates: #6259

* [TASK] improve message of delete pimcore admin user

Relates: #6259

* [TASK] add confirm dialog to static route delete operation

Relates: #6259

* [TASK] refactor confirm delete message

Relates: #6259

* [TASK] Use translations for type in delete confirmation

Relates: #6259

* [TASK] Change icon of delete objectbrick to make it consistent with delete class and delete fieldcollection

Relates: #6259

* [BUGFIX] fix objectbrick menu

* Update bundles/AdminBundle/Resources/public/js/pimcore/object/classificationstore/collectionsPanel.js


* Update bundles/AdminBundle/Resources/public/js/pimcore/object/classificationstore/groupsPanel.js


* Update bundles/AdminBundle/Resources/public/js/pimcore/object/classificationstore/groupsPanel.js


* Update bundles/AdminBundle/Resources/public/js/pimcore/settings/staticroutes.js


* Update bundles/AdminBundle/Resources/public/js/pimcore/object/classificationstore/propertiesPanel.js


* Update bundles/AdminBundle/Resources/public/js/pimcore/object/classificationstore/collectionsPanel.js


* Update bundles/AdminBundle/Resources/public/js/pimcore/settings/targeting/rules/panel.js


* Update bundles/AdminBundle/Resources/public/js/pimcore/settings/targeting/targetGroups/panel.js

+++ bundles/AdminBundle/Resources/public/js/pimcore/report/custom/panel.js
@@ -206,16 +206,20 @@ pimcore.report.custom.panel = Class.create({
     },
 
     deleteField: function (tree, record) {
-        Ext.Ajax.request({
-            url: Routing.generate('pimcore_admin_reports_customreport_delete'),
-            method: 'DELETE',
-            params: {
-                name: record.data.id
-            }
-        });
+        Ext.Msg.confirm(t('delete'), sprintf(t('delete_message_advanced'), t('portlet_customreport'), record.data.text), function (btn) {
+            if (btn == 'yes') {
+                Ext.Ajax.request({
+                    url: Routing.generate('pimcore_admin_reports_customreport_delete'),
+                    method: 'DELETE',
+                    params: {
+                        name: record.data.id
+                    }
+                });
 
-        this.getEditPanel().removeAll();
-        record.remove();
+                this.getEditPanel().removeAll();
+                record.remove();
+            }
+        }.bind(this));
     },
